+++ web/documentserver-example/csharp-mvc/Helpers/Users.cs
@@ -32,6 +32,7 @@ namespace OnlineEditorsExampleMVC.Helpers
             "The file favorite state is undefined",
             "Can create files from templates using data from the editor",
             "Can see the information about all users",
+            "Has an avatar",
             "Can submit forms"
         };
 
@@ -43,6 +44,7 @@ namespace OnlineEditorsExampleMVC.Helpers
             "This file is marked as favorite",
             "Can create new files from the editor",
             "Can see the information about users from Group2 and users who don’t belong to any group",
+            "Has an avatar",
             "Can’t submit forms"
         };
 
@@ -89,6 +91,7 @@ namespace OnlineEditorsExampleMVC.Helpers
                     null,
                     new List<string>(),
                     descr_user_1,
+                    true,
                     true
                 ),
             new User(
@@ -107,7 +110,8 @@ namespace OnlineEditorsExampleMVC.Helpers
                     true,
                     new List<string>(),
                     descr_user_2,
-                    false
+                    false,
+                    true
                 ),
             new User(
                     "uid-3",
@@ -125,6 +129,7 @@ namespace OnlineEditorsExampleMVC.Helpers
                     false,
                     new List<string>() { "copy", "download", "print" },
                     descr_user_3,
+                    false,
                     false
                 ),
             new User(
@@ -138,6 +143,7 @@ namespace OnlineEditorsExampleMVC.Helpers
                     null,
                     new List<string>() { "protect" },
                     descr_user_0,
+                    false,
                     false
                 )
         };
@@ -176,6 +182,24 @@ namespace OnlineEditorsExampleMVC.Helpers
             return usersData;
         }
 
+        public static List<Dictionary<string, object>> getUsersInfo(string id)
+        {
+            List<Dictionary<string, object>> usersData = new List<Dictionary<string, object>>();
+            if (id != "uid-0") {
+                foreach (User user in users)
+                {
+                    usersData.Add(new Dictionary<string, object>()
+                        {
+                            {"id", user.id},
+                            {"name", user.name },
+                            {"email", user.email },
+                            {"image", user.avatar ? DocManagerHelper.GetServerUrl(false) + "/Content/images/" + user.id + ".png" : null}
+                        });
+                }
+            }
+            return usersData;
+        }
+
         // get a list of users with their names and emails for protect
         public static List<Dictionary<string, object>> getUsersForProtect(string id)
         {
@@ -209,8 +233,9 @@ namespace OnlineEditorsExampleMVC.Helpers
         public List<string> descriptions;
         public bool templates;
         public List<string> userInfoGroups;
+        public bool avatar;
 
-        public User(string id, string name, string email, string group, List<string> reviewGroups, Dictionary<string, object> commentGroups, List<string> userInfoGroups, bool? favorite, List<string> deniedPermissions, List<string> descriptions, bool templates)
+        public User(string id, string name, string email, string group, List<string> reviewGroups, Dictionary<string, object> commentGroups, List<string> userInfoGroups, bool? favorite, List<string> deniedPermissions, List<string> descriptions, bool templates, bool avatar)
         {
             this.id = id;
             this.name = name;
@@ -223,6 +248,7 @@ namespace OnlineEditorsExampleMVC.Helpers
             this.descriptions = descriptions;
             this.templates = templates;
             this.userInfoGroups = userInfoGroups;
+            this.avatar = avatar;
         }
     }
 }