+++ web/documentserver-example/java/src/main/java/controllers/EditorServlet.java
@@ -108,6 +108,8 @@ public class EditorServlet extends HttpServlet {
         List<Map<String, Object>> usersForMentions = Users.getUsersForMentions(user.getId());
         List<Map<String, Object>> usersForProtect = Users.getUsersForProtect(user.getId());
 
+        List<Map<String, Object>> usersInfo = Users.getUsersInfo(user.getId());
+
         // check if the document token is enabled
         if (DocumentManager.tokenEnabled()) {
             file.buildToken();  // generate document token
@@ -132,6 +134,7 @@ public class EditorServlet extends HttpServlet {
         request.setAttribute("dataSpreadsheet", gson.toJson(dataSpreadsheet));
         request.setAttribute("usersForMentions", !user.getId()
                 .equals("uid-0") ? gson.toJson(usersForMentions) : null);
+        request.setAttribute("usersInfo", gson.toJson(usersInfo));
         request.setAttribute("usersForProtect", !user.getId()
                 .equals("uid-0") ? gson.toJson(usersForProtect) : null);
         request.getRequestDispatcher("editor.jsp").forward(request, response);