+++ web/documentserver-example/php/src/helpers/ExampleUsers.php
@@ -38,6 +38,7 @@ final class ExampleUsers
             "The file favorite state is undefined",
             "Can create files from templates using data from the editor",
             "Can see the information about all users",
+            "Has an avatar",
             "Can submit forms"
         ];
         $this->user2Description = [
@@ -48,6 +49,7 @@ final class ExampleUsers
             "This file is marked as favorite",
             "Can create new files from the editor",
             "Can see the information about users from Group2 and users who don’t belong to any group",
+            "Has an avatar",
             "Can’t submit forms"
         ];
         $this->user3Description = [
@@ -88,6 +90,7 @@ final class ExampleUsers
                 null,
                 [],
                 $this->user1Description,
+                true,
                 true
             ),
             new Users(
@@ -105,7 +108,8 @@ final class ExampleUsers
                 true,
                 [],
                 $this->user2Description,
-                false
+                false,
+                true
             ),
             new Users(
                 "uid-3",
@@ -122,6 +126,7 @@ final class ExampleUsers
                 false,
                 ["copy", "download", "print"],
                 $this->user3Description,
+                false,
                 false
             ),
             new Users(
@@ -135,6 +140,7 @@ final class ExampleUsers
                 null,
                 ["protect"],
                 $this->user0Description,
+                false,
                 false
             ),
         ];